+++ src/movies/movies.service.ts
@@ -1,10 +1,11 @@
-import { ConflictException, Injectable, NotFoundException, Res } from '@nestjs/common';
+import { ConflictException, ForbiddenException, Injectable, NotFoundException, Res, UnauthorizedException } from '@nestjs/common';
 import { CreateMovieDto } from './dto/create-movie.dto';
 import { UpdateMovieDto } from './dto/update-movie.dto';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { Movie } from './entities/movie.entity';
 import { Category } from 'src/categories/entities/category.entity';
+import { UserRoleEnum } from 'src/users/entities/user.entity';
 
 @Injectable()
 export class MoviesService {
@@ -29,13 +30,16 @@ export class MoviesService {
       }
     });
     if (!movie) {
-      console.log(movie)
       throw new NotFoundException(`La película con el Id ${id} no existe`);
     }
     return movie;
   }
 
-  async create(createMovieDto: CreateMovieDto): Promise<Movie> {
+  async create(createMovieDto: CreateMovieDto, user_id: number, role: UserRoleEnum): Promise<Movie> {
+    if (role != UserRoleEnum.ADMIN) {
+      throw new ForbiddenException('Usted no está autorizado para crear películas');
+    }
+
     const existsMovie = await this.moviesRepository.exists({
       where: { 
         title: createMovieDto.title
@@ -52,7 +56,7 @@ export class MoviesService {
     if (!existsCategory) {
       throw new ConflictException('La categoría no existe');
     }
-    return this.moviesRepository.save({ ...createMovieDto });
+    return this.moviesRepository.save({ ...createMovieDto, createdBy: user_id });
   }
 
   async findAll(page = 1, limit = 10, relations = false): Promise<{ data: Movie[]; total: number; page: number; limit: number }>  {
@@ -87,7 +91,11 @@ export class MoviesService {
     return this.findOneOrFail(id, relations);
   }
 
-  async update(id: number, updateMovieDto: UpdateMovieDto): Promise<Movie> {
+  async update(id: number, updateMovieDto: UpdateMovieDto, user_id: number, role: UserRoleEnum): Promise<Movie> {
+    if (role != UserRoleEnum.ADMIN) {
+      throw new ForbiddenException('Usted no está autorizado para modificar películas');
+    }
+
     const movie = await this.findOneOrFail(id);
 
     if (updateMovieDto.title != null) {
@@ -119,11 +127,16 @@ export class MoviesService {
       }
       movie.categoryId = updateMovieDto.categoryId;
     }
+    movie.updatedBy = user_id;
 
     return this.moviesRepository.save(movie);
   }
 
-  async remove(id: number) {
+  async remove(id: number, role: UserRoleEnum) {
+    if (role != UserRoleEnum.ADMIN) {
+      throw new ForbiddenException('Usted no está autorizado para eliminar películas');
+    }
+
     const movie = await this.findOneOrFail(id);
 
     return this.moviesRepository.delete(id);
